requirements.txt
asgiref==3.8.1
Django==4.2.16
djangorestframework==3.15.2
drf-yasg==1.21.7
Faker==28.4.1
inflection==0.5.1
packaging==24.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
